+++ project/src/store/root-action.ts
@@ -1,0 +1,5 @@
+import {createAction} from '@reduxjs/toolkit';
+
+import {AppRoute} from 'components/app/const';
+
+export const redirectToRoute = createAction<AppRoute>('offers/redirectToRoute');
